refactor(onboard): extract ollama proxy into ollama-proxy.ts, add isProxyHealthy (#767)
Extract the entire Ollama auth proxy subsystem (token persistence, PID
tracking, process spawn/kill, health probing) from onboard.ts into a
dedicated ollama-proxy.ts module.

Add isProxyHealthy() which performs an actual HTTP probe against the
proxy endpoint (not just a PID check), addressing CodeRabbit's review
feedback about wedged/stale proxy processes.

Simplify the setupInference fallback to use isProxyHealthy() instead of
inline PID + host check logic.

- Reduces onboard.ts by 161 lines (7910 → 7749)
- New ollama-proxy.ts: 223 lines
- Advances #767 (split onboarding provider logic)
